FolderPage: Replaced pykdeuic4 generated code with cleaned up original ui file. Merged with UI_FolderPage class and moved everything to folderpage.py
