+++ lib/autokey/qtui/folderpage.py
@@ -1,53 +1,83 @@
-#!/usr/bin/env python
-# coding=UTF-8
-#
-# Generated by pykdeuic4 from folderpage.ui on Sun Mar  4 11:39:39 2012
-#
-# WARNING! All changes to this file will be lost.
-from PyKDE4 import kdecore
-from PyKDE4 import kdeui
-from PyQt4 import QtCore, QtGui
+# Copyright (C) 2011 Chris Dekter
+# Copyright (C) 2018 Thomas Hess <thomas.hess@udo.edu>
 
-try:
-    _fromUtf8 = QtCore.QString.fromUtf8
-except AttributeError:
-    _fromUtf8 = lambda s: s
+# This program is free software: you can redistribute it and/or modify
+# it under the terms of the GNU General Public License as published by
+# the Free Software Foundation, either version 3 of the License, or
+# (at your option) any later version.
 
-class Ui_FolderPage(object):
-    def setupUi(self, FolderPage):
-        FolderPage.setObjectName(_fromUtf8("FolderPage"))
-        FolderPage.resize(568, 530)
-        self.verticalLayout_2 = QtGui.QVBoxLayout(FolderPage)
-        self.verticalLayout_2.setObjectName(_fromUtf8("verticalLayout_2"))
-        self.urlLabel = KUrlLabel(FolderPage)
-        self.urlLabel.setAlignment(QtCore.Qt.AlignCenter)
-        self.urlLabel.setObjectName(_fromUtf8("urlLabel"))
-        self.verticalLayout_2.addWidget(self.urlLabel)
-        self.settingsGroupBox = QtGui.QGroupBox(FolderPage)
-        self.settingsGroupBox.setObjectName(_fromUtf8("settingsGroupBox"))
-        self.verticalLayout = QtGui.QVBoxLayout(self.settingsGroupBox)
-        self.verticalLayout.setObjectName(_fromUtf8("verticalLayout"))
-        self.showInTrayCheckbox = QtGui.QCheckBox(self.settingsGroupBox)
-        self.showInTrayCheckbox.setObjectName(_fromUtf8("showInTrayCheckbox"))
-        self.verticalLayout.addWidget(self.showInTrayCheckbox)
-        self.kseparator = KSeparator(self.settingsGroupBox)
-        self.kseparator.setObjectName(_fromUtf8("kseparator"))
-        self.verticalLayout.addWidget(self.kseparator)
-        self.settingsWidget = SettingsWidget(self.settingsGroupBox)
-        self.settingsWidget.setObjectName(_fromUtf8("settingsWidget"))
-        self.verticalLayout.addWidget(self.settingsWidget)
-        self.verticalLayout_2.addWidget(self.settingsGroupBox)
-        spacerItem = QtGui.QSpacerItem(20, 40, QtGui.QSizePolicy.Minimum, QtGui.QSizePolicy.Expanding)
-        self.verticalLayout_2.addItem(spacerItem)
+# This program is distributed in the hope that it will be useful,
+# but WITHOUT ANY WARRANTY; without even the implied warranty of
+# MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+# GNU General Public License for more details.
 
-        self.retranslateUi(FolderPage)
-        QtCore.QMetaObject.connectSlotsByName(FolderPage)
+# You should have received a copy of the GNU General Public License
+# along with this program. If not, see <http://www.gnu.org/licenses/>.
+import subprocess
 
-    def retranslateUi(self, FolderPage):
-        FolderPage.setWindowTitle(kdecore.i18n(_fromUtf8("Form")))
-        self.urlLabel.setTipText(kdecore.i18n(_fromUtf8("Open the folder in the default file manager")))
-        self.settingsGroupBox.setTitle(kdecore.i18n(_fromUtf8("Folder Settings")))
-        self.showInTrayCheckbox.setText(kdecore.i18n(_fromUtf8("Show in notification icon menu")))
+from PyQt4.QtGui import QMessageBox
 
-from PyKDE4.kdeui import KUrlLabel, KSeparator
-from .settingswidget import SettingsWidget
+from .common import inherits_from_ui_file_with_name, set_url_label
+
+from .. import configmanager as cm
+
+PROBLEM_MSG_PRIMARY = "Some problems were found"
+PROBLEM_MSG_SECONDARY = "{}\n\nYour changes have not been saved."
+
+
+class FolderPage(*inherits_from_ui_file_with_name("folderpage")):
+
+    def __init__(self):
+        super(FolderPage, self).__init__()
+        self.setupUi(self)
+
+    def load(self, folder):
+        self.currentFolder = folder
+        self.showInTrayCheckbox.setChecked(folder.showInTrayMenu)
+        self.settingsWidget.load(folder)
+
+        if self.is_new_item():
+            self.urlLabel.setEnabled(False)
+            self.urlLabel.setText("(Unsaved)")  # TODO: i18n
+        else:
+            set_url_label(self.urlLabel, self.currentFolder.path)
+
+    def save(self):
+        self.currentFolder.showInTrayMenu = self.showInTrayCheckbox.isChecked()
+        self.settingsWidget.save()
+        self.currentFolder.persist()
+        set_url_label(self.urlLabel, self.currentFolder.path)
+
+        return not self.currentFolder.path.startswith(cm.CONFIG_DEFAULT_FOLDER)
+
+    def set_item_title(self, title):
+        self.currentFolder.title = title
+
+    def rebuild_item_path(self):
+        self.currentFolder.rebuild_path()
+
+    def is_new_item(self):
+        return self.currentFolder.path is None
+
+    def reset(self):
+        self.load(self.currentFolder)
+
+    def validate(self):
+        # Check settings
+        errors = self.settingsWidget.validate()
+
+        if errors:
+            msg = PROBLEM_MSG_SECONDARY.format('\n'.join([str(e) for e in errors]))
+            QMessageBox.critical(self.topLevelWidget(), PROBLEM_MSG_PRIMARY, msg)
+
+        return not bool(errors)
+
+    def set_dirty(self):
+        self.topLevelWidget().set_dirty()
+
+    # --- Signal handlers
+    def on_showInTrayCheckbox_stateChanged(self, state):
+        self.set_dirty()
+
+    def on_urlLabel_leftClickedUrl(self, url=None):
+        if url: subprocess.Popen(["/usr/bin/xdg-open", url])
