+++ lib/autokey/qtui/resources/ui/folderpage.ui
@@ -11,16 +11,28 @@
    </rect>
   </property>
   <property name="windowTitle">
-   <string>Form</string>
+   <string>FolderPage</string>
   </property>
   <layout class="QVBoxLayout" name="verticalLayout_2">
    <item>
-    <widget class="KUrlLabel" name="urlLabel">
+    <widget class="QLabel" name="urlLabel">
+     <property name="toolTip">
+      <string>Open the folder in the default file manager</string>
+     </property>
+     <property name="textFormat">
+      <enum>Qt::RichText</enum>
+     </property>
      <property name="alignment">
       <set>Qt::AlignCenter</set>
      </property>
-     <property name="tipText">
-      <string>Open the folder in the default file manager</string>
+     <property name="openExternalLinks">
+      <bool>true</bool>
+     </property>
+     <property name="textInteractionFlags">
+      <set>Qt::TextBrowserInteraction</set>
+     </property>
+     <property name="tipText" stdset="0">
+      <string/>
      </property>
     </widget>
    </item>
@@ -38,7 +50,11 @@
        </widget>
       </item>
       <item>
-       <widget class="KSeparator" name="kseparator"/>
+       <widget class="Line" name="line">
+        <property name="orientation">
+         <enum>Qt::Horizontal</enum>
+        </property>
+       </widget>
       </item>
       <item>
        <widget class="SettingsWidget" name="settingsWidget" native="true"/>
@@ -63,19 +79,9 @@
  </widget>
  <customwidgets>
   <customwidget>
-   <class>KSeparator</class>
-   <extends>QFrame</extends>
-   <header>kseparator.h</header>
-  </customwidget>
-  <customwidget>
-   <class>KUrlLabel</class>
-   <extends>QLabel</extends>
-   <header>kurllabel.h</header>
-  </customwidget>
-  <customwidget>
    <class>SettingsWidget</class>
    <extends>QWidget</extends>
-   <header>configwindow.h</header>
+   <header>autokey.qtui.settingswidget</header>
    <container>1</container>
   </customwidget>
  </customwidgets>
